+++ src/main/java/nl/clockwork/ebms/dao/hsqldb/EbMSDAOImpl.java
@@ -33,7 +33,7 @@ public class EbMSDAOImpl extends AbstractEbMSDAO
 	}
 
 	@Override
-	public String getMessageIdsQuery(String messageContextFilter, EbMSMessageStatus status, int maxNr)
+	public String getMessageIdsQuery(String messageContextFilter, EbMSMessageStatus status, long maxNr)
 	{
 		return "select message_id" +
 		" from ebms_message" +
